+++ mex/private/mexIgraphRandGame.c
@@ -139,10 +139,11 @@ static igraph_error_t mxIgraph_watts_strogatz_i(mxArray const* opts,
   igraph_integer_t n_neighbors = mxIgraphGetInteger(opts, "radius");
   igraph_real_t probability = mxIgraphGetReal(opts, "probability");
   igraph_bool_t loops = mxIgraphGetBool(opts, "loops");
+  igraph_bool_t multiple = mxIgraphGetBool(opts, "multiple");
   igraph_error_t errcode;
 
   errcode = igraph_watts_strogatz_game(graph, dim, size, n_neighbors,
-                                       probability, loops, false);
+                                       probability, loops, multiple);
 
   return errcode;
 }
@@ -198,9 +199,10 @@ static igraph_error_t mxIgraph_k_regular_i(mxArray const* opts,
   igraph_integer_t n_nodes = mxIgraphGetInteger(opts, "nNodes");
   igraph_integer_t degree = mxIgraphGetInteger(opts, "degree");
   igraph_bool_t directed = mxIgraphGetBool(opts, "isdirected");
+  igraph_bool_t multiple = mxIgraphGetBool(opts, "multiple");
   igraph_error_t error;
 
-  error = igraph_k_regular_game(graph, n_nodes, degree, directed, false);
+  error = igraph_k_regular_game(graph, n_nodes, degree, directed, multiple);
 
   return error;
 }
@@ -213,6 +215,7 @@ static igraph_error_t mxIgraph_static_fitness_i(mxArray const* opts,
   igraph_vector_t fitness_in;
   igraph_vector_t* fitness_in_ptr = NULL;
   igraph_bool_t loops = mxIgraphGetBool(opts, "loops");
+  igraph_bool_t multiple = mxIgraphGetBool(opts, "multiple");
   igraph_error_t error;
 
   mxIgraphGetVector(opts, "outFitness", &fitness_out, false);
@@ -223,7 +226,7 @@ static igraph_error_t mxIgraph_static_fitness_i(mxArray const* opts,
   }
 
   error = igraph_static_fitness_game(graph, n_edges, &fitness_out,
-                                     fitness_in_ptr, loops, false);
+                                     fitness_in_ptr, loops, multiple);
 
   igraph_vector_destroy(&fitness_out);
   igraph_vector_destroy(&fitness_in);
@@ -239,12 +242,13 @@ static igraph_error_t mxIgraph_static_power_law_i(mxArray const
   igraph_real_t exponent_out = mxIgraphGetReal(opts, "outExponent");
   igraph_real_t exponent_in = mxIgraphGetReal(opts, "inExponent");
   igraph_bool_t loops = mxIgraphGetBool(opts, "loops");
+  igraph_bool_t multiple = mxIgraphGetBool(opts, "multiple");
   igraph_bool_t finite_size_correction = mxIgraphGetBool(opts,
                                          "finiteSizeCorrection");
   igraph_error_t error;
 
   error = igraph_static_power_law_game(graph, n_nodes, n_edges, exponent_out,
-                                       exponent_in, loops, false,
+                                       exponent_in, loops, multiple,
                                        finite_size_correction);
 
   return error;