+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Graph `summary` function.
 - A main igraph help entry used to describe common function arguments.
 - Reindex membership.
+- Allow `multiple` option for algorithms that allow when using MATLAB graphs.
 
 ### Changed